improve glibc threading interception
previously we relied on preload to inject user-allocated stacks for
pthread_create, freeing them on pthread_join. To support pthread_detach,
we now replace the preload mechanism with direct modifications to glibc:
glibc calls into junction/caladan when pthreads are created to allocate a
stack. Before a detached thread exits, it registers its stack with the runtime
so that the runtime can free it itself.