fix vdso functions
Although we pass a pointer to the vdso in auxv, some binaries do not use it.
Junction can rely on vdso to implement some syscalls, but using the glibc
wrappers for these functions requires switching fsbase and reading error codes
from errno. Instead, install pointers to the vdso functions directly into the
syscall table for binaries that don't use vdso.