updates to interruptible wait
The previous code for interruptible_wake() used a buggy check to determine
if the parked thread had called prepare_interruptible() or not. If a signal
was delivered to a thread that was parked and uninterruptible, upon wake
the waker would decrement the wait counter leading to a missed signal.

The fix uses a high bit in the interrupt_state to encode that a
thread is prepared.