uintr: use xsaveopt when possible
we reserve an area at the bottom of the per-uthread syscall stack for a full
xsave buffer. when performing an xsave, we check if our most recent xrstor
on this cpu core was performed with the same thread/buffer. If so, xsaveopt
can be used. dedicating a buffer for each thread also reduces copying of the
xsave area when delivering signals.