signal: extra fault checks for snapshotting
An instance that will be snapshotted cannot contain any pointers to Junction
code in its metadata. This can happen when a signal is delivered just as a
syscall is beginning. This commit adds some checks to determine if this has
occurred so the signal handler can react accordingly. These checks are only
used if a runtime is started with the snapshot_enabled flag (or the -S flag).