mm: tracer fixes
We ran into issues with the tracer's page fault handler needing to lock the
memory map while another thread is already holding the lock. This is a problem
when (a) the fault occurred while preemption was disabled, and the fault
function must spin to acquire the lock or (b) when the fault occurred in
the function holding the lock (eg the stack itself causes the fault).

This commit fixes problem (a) by switching to using a spin lock with
preemption disabled for MM modifications when tracing is enabled. Problem (b)
is fixed by forcing a stack switch to the system call stack when tracing is
enabled and the spin lock must be acquired.

This commit also includes fixes to the tracer to handle VMAs that modify their
permissions during tracing properly.