uintr: fix bug in stackswitch mode
stack-switching syscall entry places some data on the stack before switching
the rsp but the uintr handler may also place parts of the trapframe on the
syscall stack when the interrupted rsp is not on the syscall stack.
the fix is to consider a redzone area when pushing data to the syscall
stack.

Thanks to @hky1999 for their work on this issue.